Refactor out numpy FFTKits and ResampKits.
