+++ sfft/SpaceSFFTNumpyFlow.py
@@ -1,286 +1,18 @@
 # IMPORTS Standard
-import math
 import numpy as np
 
 # IMPORTS Internal
 from sfft.sfftcore.SFFTSubtract import GeneralSFFTSubtract
 from sfft.sfftcore.SFFTConfigure import SingleSFFTConfigure
 from sfft.utils.PatternRotationCalculator import PatternRotation_Calculator
-from sfft.utils.ReadWCS import Read_WCS
+from sfft.utils.NumpyFFTKits import Numpy_FFTKits
+from sfft.utils.NumpyResampKits import Numpy_Resampling, Numpy_ZoomRotate
 from sfft.utils.DeCorrelationCalculator import DeCorrelation_Calculator, KERNEL_CSZ, KERNEL_CSZ_INV
 from sfft.utils.SkyLevelEstimator import SkyLevel_Estimator
 from sfft.utils.SFFTSolutionReader import Realize_MatchingKernel
 
 __last_update__ = "2026-05-20"
 __author__ = "Lei Hu <leihu@andrew.cmu.edu>, Michael Wood-Vasey <wmwv@pitt.edu>, and VS Code Copilot"
-
-
-class Numpy_FFTKits:
-    @staticmethod
-    def KERNEL_CSZ(KERNEL, NX_IMG, NY_IMG, NORMALIZE_KERNEL=False):
-        """Circular shift the kernel and extend to the target size."""
-        N0, N1 = NX_IMG, NY_IMG
-        L0, L1 = KERNEL.shape
-        W0, W1 = (L0 - 1) // 2, (L1 - 1) // 2
-        assert L0 % 2 == 1 and L1 % 2 == 1
-
-        if NORMALIZE_KERNEL:
-            KERNEL_TZP = np.pad(
-                KERNEL / np.sum(KERNEL), pad_width=((0, N0 - L0), (0, N1 - L1)), mode="constant", constant_values=0.0
-            )
-        else:
-            KERNEL_TZP = np.pad(KERNEL, pad_width=((0, N0 - L0), (0, N1 - L1)), mode="constant", constant_values=0.0)
-        KIMG_CSZ = np.roll(np.roll(KERNEL_TZP, -W0, axis=0), -W1, axis=1)
-        return KIMG_CSZ
-
-    @staticmethod
-    def KERNEL_CSZ_INV(KIMG, NX_KERN, NY_KERN, VERBOSE_LEVEL=2):
-        """Inverse circular shift the kernel and truncate to the target size."""
-        L0, L1 = NX_KERN, NY_KERN
-        W0, W1 = (L0 - 1) // 2, (L1 - 1) // 2
-        assert L0 % 2 == 1 and L1 % 2 == 1
-
-        KIMG_iCSZ = np.roll(np.roll(KIMG, W1, axis=1), W0, axis=0)
-        KERNEL = KIMG_iCSZ[:L0, :L1]
-        if VERBOSE_LEVEL in [1, 2]:
-            LOSE_RATIO = 1.0 - np.sum(np.abs(KERNEL)) / np.sum(np.abs(KIMG_iCSZ))
-            _report_message = "Kernel Truncation Loses APE = [%.4f %s]" % (LOSE_RATIO * 100, "%")
-            print("MeLOn CheckPoint: %s " % _report_message)
-        return KERNEL
-
-    @staticmethod
-    def FFT_CONVOLVE(
-        PixA_Inp,
-        KERNEL,
-        PAD_FILL_VALUE=0.0,
-        NAN_FILL_VALUE=0.0,
-        NORMALIZE_KERNEL=False,
-        FORCE_OUTPUT_C_CONTIGUOUS=False,
-        FFT_BACKEND="Numpy",
-    ):
-        """FFT convolution with zero padding."""
-        N0, N1 = PixA_Inp.shape
-        L0, L1 = KERNEL.shape
-        assert L0 % 2 == 1 and L1 % 2 == 1
-
-        W0, W1 = (L0 - 1) // 2, (L1 - 1) // 2
-        NX_IMG, NY_IMG = N0 + 2 * W0, N1 + 2 * W1
-
-        PixA_EInp = np.pad(PixA_Inp, pad_width=((W0, W0), (W1, W1)), mode="constant", constant_values=PAD_FILL_VALUE)
-
-        if NAN_FILL_VALUE is not None:
-            PixA_EInp[np.isnan(PixA_EInp)] = NAN_FILL_VALUE
-
-        KIMG_CSZ = Numpy_FFTKits.KERNEL_CSZ(
-            KERNEL=KERNEL, NX_IMG=NX_IMG, NY_IMG=NY_IMG, NORMALIZE_KERNEL=NORMALIZE_KERNEL
-        )
-
-        PixA_Out = np.fft.ifft2(np.fft.fft2(PixA_EInp) * np.fft.fft2(KIMG_CSZ)).real[W0:-W0, W1:-W1]
-
-        if FORCE_OUTPUT_C_CONTIGUOUS and not PixA_Out.flags["C_CONTIGUOUS"]:
-            PixA_Out = np.ascontiguousarray(PixA_Out)
-        return PixA_Out
-
-
-class Numpy_Resampling:
-    def __init__(self, RESAMP_METHOD="BILINEAR", VERBOSE_LEVEL=2):
-        self.RESAMP_METHOD = RESAMP_METHOD
-        self.VERBOSE_LEVEL = VERBOSE_LEVEL
-
-    def resamp_projection_astropy(self, hdr_obj, hdr_targ):
-        """Project the target pixel centers to the object frame using Astropy WCS."""
-        NTX = int(hdr_targ["NAXIS1"])
-        NTY = int(hdr_targ["NAXIS2"])
-
-        XX_targ, YY_targ = np.meshgrid(
-            np.arange(1, NTX + 1, dtype=np.float64), np.arange(1, NTY + 1, dtype=np.float64), indexing="ij"
-        )
-        XY_targ = np.vstack((XX_targ.ravel(), YY_targ.ravel())).T
-
-        w_targ = Read_WCS.RW(hdr_targ, VERBOSE_LEVEL=self.VERBOSE_LEVEL)
-        w_obj = Read_WCS.RW(hdr_obj, VERBOSE_LEVEL=self.VERBOSE_LEVEL)
-
-        XY_world = w_targ.all_pix2world(XY_targ, 1)
-        XY_proj = w_obj.all_world2pix(XY_world, 1)
-
-        XX_proj = XY_proj[:, 0].reshape((NTX, NTY))
-        YY_proj = XY_proj[:, 1].reshape((NTX, NTY))
-        return XX_proj, YY_proj
-
-    def frame_extension(self, XX_proj, YY_proj, PixA_obj, PAD_FILL_VALUE=0.0, NAN_FILL_VALUE=0.0):
-        """Extend the object image to cover the projected coordinates."""
-        NTX, NTY = XX_proj.shape
-        NOX, NOY = PixA_obj.shape
-
-        if self.RESAMP_METHOD == "BILINEAR":
-            KERHW = (1, 1)
-        else:
-            raise ValueError("Unsupported RESAMP_METHOD '%s'" % self.RESAMP_METHOD)
-
-        RMIN = math.floor(np.min(XX_proj)) - 1 - KERHW[0]
-        RMAX = math.floor(np.max(XX_proj)) - 1 + KERHW[0]
-        RPAD = (-min(RMIN, 0), max(RMAX - (NOX - 1), 0))
-
-        CMIN = math.floor(np.min(YY_proj)) - 1 - KERHW[1]
-        CMAX = math.floor(np.max(YY_proj)) - 1 + KERHW[1]
-        CPAD = (-min(CMIN, 0), max(CMAX - (NOY - 1), 0))
-
-        PAD_WIDTH = (RPAD, CPAD)
-        PixA_Eobj = np.pad(PixA_obj, PAD_WIDTH, mode="constant", constant_values=PAD_FILL_VALUE)
-        if NAN_FILL_VALUE is not None:
-            PixA_Eobj[np.isnan(PixA_Eobj)] = NAN_FILL_VALUE
-
-        XX_Eproj = XX_proj + PAD_WIDTH[0][0]
-        YY_Eproj = YY_proj + PAD_WIDTH[1][0]
-
-        EProjDict = {
-            "NTX": NTX,
-            "NTY": NTY,
-            "NOX": NOX,
-            "NOY": NOY,
-            "NEOX": PixA_Eobj.shape[0],
-            "NEOY": PixA_Eobj.shape[1],
-            "KERHW": KERHW,
-            "XX_Eproj": XX_Eproj,
-            "YY_Eproj": YY_Eproj,
-        }
-        return PixA_Eobj, EProjDict
-
-    def resampling(
-        self,
-        PixA_Eobj,
-        EProjDict,
-        PIXEL_SCALE_FACTOR=1.0,
-        CUDA_COMPILER="nvrtc",
-        THREAD_PER_BLOCK=8,
-        USE_SHARED_MEMORY=False,
-    ):
-        """Bilinear resampling using NumPy."""
-        NTX = EProjDict["NTX"]
-        NTY = EProjDict["NTY"]
-        XX_Eproj = EProjDict["XX_Eproj"]
-        YY_Eproj = EProjDict["YY_Eproj"]
-
-        PixA_resamp = np.zeros((NTX, NTY), dtype=np.float64)
-
-        if self.RESAMP_METHOD == "BILINEAR":
-            if USE_SHARED_MEMORY and self.VERBOSE_LEVEL in [1, 2]:
-                print("MeLOn WARNING: Shared memory is not implemented for NumPy bilinear resampling!")
-
-            R1 = np.floor(XX_Eproj).astype(np.int64) - 1
-            C1 = np.floor(YY_Eproj).astype(np.int64) - 1
-            R2 = R1 + 1
-            C2 = C1 + 1
-
-            DX = XX_Eproj - np.floor(XX_Eproj)
-            DY = YY_Eproj - np.floor(YY_Eproj)
-
-            W11 = (1.0 - DX) * (1.0 - DY)
-            W12 = (1.0 - DX) * DY
-            W21 = DX * (1.0 - DY)
-            W22 = DX * DY
-
-            PixA_resamp = (
-                W11 * PixA_Eobj[R1, C1] + W12 * PixA_Eobj[R1, C2] + W21 * PixA_Eobj[R2, C1] + W22 * PixA_Eobj[R2, C2]
-            )
-            PixA_resamp *= PIXEL_SCALE_FACTOR
-        else:
-            raise ValueError("Unsupported RESAMP_METHOD '%s'" % self.RESAMP_METHOD)
-
-        return PixA_resamp
-
-
-class Numpy_ZoomRotate:
-    @staticmethod
-    def CZR(
-        PixA_obj,
-        ZOOM_SCALE_X=1.0,
-        ZOOM_SCALE_Y=1.0,
-        OUTSIZE_PARITY_X="UNCHANGED",
-        OUTSIZE_PARITY_Y="UNCHANGED",
-        PATTERN_ROTATE_ANGLE=0.0,
-        RESAMP_METHOD="BILINEAR",
-        PAD_FILL_VALUE=0.0,
-        NAN_FILL_VALUE=0.0,
-        CUDA_COMPILER="nvrtc",
-        THREAD_PER_BLOCK=8,
-        USE_SHARED_MEMORY=False,
-        VERBOSE_LEVEL=2,
-    ):
-        """Zoom and rotate an image using NumPy."""
-        assert ZOOM_SCALE_X > 0.0
-        assert ZOOM_SCALE_Y > 0.0
-        assert 0.0 <= PATTERN_ROTATE_ANGLE < 360.0
-
-        NAXIS1_ORI, NAXIS2_ORI = PixA_obj.shape
-
-        def NPIX_ORI2ZOOMED(NPIX_ORI, ZOOM_SCALE, OUTSIZE_PARITY):
-            if OUTSIZE_PARITY == "UNCHANGED":
-                OUTSIZE_UPARITY = "EVEN" if NPIX_ORI % 2 == 0 else "ODD"
-            elif OUTSIZE_PARITY == "ODD":
-                OUTSIZE_UPARITY = "ODD"
-            elif OUTSIZE_PARITY == "EVEN":
-                OUTSIZE_UPARITY = "EVEN"
-            else:
-                raise ValueError("Invalid OUTSIZE_PARITY value: %s" % OUTSIZE_PARITY)
-
-            HALFWIDTH_ORI = NPIX_ORI / 2.0
-            if OUTSIZE_UPARITY == "EVEN":
-                return 2 * math.ceil(HALFWIDTH_ORI / ZOOM_SCALE)
-            return 2 * math.ceil((HALFWIDTH_ORI - ZOOM_SCALE / 2.0) / ZOOM_SCALE) + 1
-
-        NAXIS1_ZOOMED = NPIX_ORI2ZOOMED(NPIX_ORI=NAXIS1_ORI, ZOOM_SCALE=ZOOM_SCALE_X, OUTSIZE_PARITY=OUTSIZE_PARITY_X)
-        NAXIS2_ZOOMED = NPIX_ORI2ZOOMED(NPIX_ORI=NAXIS2_ORI, ZOOM_SCALE=ZOOM_SCALE_Y, OUTSIZE_PARITY=OUTSIZE_PARITY_Y)
-
-        def BACKWARD_TRANSFORM(X_ROTATED, Y_ROTATED):
-            CRPIX1_ZOOMED = 0.5 + NAXIS1_ZOOMED / 2.0
-            CRPIX2_ZOOMED = 0.5 + NAXIS2_ZOOMED / 2.0
-
-            ALPHA_INV = np.deg2rad(-PATTERN_ROTATE_ANGLE)
-            ROTMAT_INV = np.array(
-                [[np.cos(ALPHA_INV), -np.sin(ALPHA_INV)], [np.sin(ALPHA_INV), np.cos(ALPHA_INV)]], dtype=np.float64
-            )
-
-            COORD_ROTATED = np.vstack((X_ROTATED - CRPIX1_ZOOMED, Y_ROTATED - CRPIX2_ZOOMED))
-            COORD_ZOOMED = ROTMAT_INV.dot(COORD_ROTATED)
-            COORD_ZOOMED[0, :] += CRPIX1_ZOOMED
-            COORD_ZOOMED[1, :] += CRPIX2_ZOOMED
-
-            OFFSET_X = ZOOM_SCALE_X * (0.5 + NAXIS1_ZOOMED / 2.0) - (0.5 + NAXIS1_ORI / 2.0)
-            OFFSET_Y = ZOOM_SCALE_Y * (0.5 + NAXIS2_ZOOMED / 2.0) - (0.5 + NAXIS2_ORI / 2.0)
-
-            X_ORI = COORD_ZOOMED[0, :] * ZOOM_SCALE_X - OFFSET_X
-            Y_ORI = COORD_ZOOMED[1, :] * ZOOM_SCALE_Y - OFFSET_Y
-            return X_ORI, Y_ORI
-
-        XX_targ, YY_targ = np.meshgrid(
-            np.arange(1, NAXIS1_ZOOMED + 1, dtype=np.float64),
-            np.arange(1, NAXIS2_ZOOMED + 1, dtype=np.float64),
-            indexing="ij",
-        )
-
-        X_ORI, Y_ORI = BACKWARD_TRANSFORM(X_ROTATED=XX_targ.ravel(), Y_ROTATED=YY_targ.ravel())
-        XX_proj = X_ORI.reshape((NAXIS1_ZOOMED, NAXIS2_ZOOMED))
-        YY_proj = Y_ORI.reshape((NAXIS1_ZOOMED, NAXIS2_ZOOMED))
-
-        GWR = Numpy_Resampling(RESAMP_METHOD=RESAMP_METHOD, VERBOSE_LEVEL=VERBOSE_LEVEL)
-        PixA_Eobj, EProjDict = GWR.frame_extension(
-            XX_proj=XX_proj,
-            YY_proj=YY_proj,
-            PixA_obj=PixA_obj,
-            PAD_FILL_VALUE=PAD_FILL_VALUE,
-            NAN_FILL_VALUE=NAN_FILL_VALUE,
-        )
-
-        PIXEL_SCALE_FACTOR = ZOOM_SCALE_X * ZOOM_SCALE_Y
-        PixA_resamp = GWR.resampling(
-            PixA_Eobj=PixA_Eobj,
-            EProjDict=EProjDict,
-            PIXEL_SCALE_FACTOR=PIXEL_SCALE_FACTOR,
-            USE_SHARED_MEMORY=USE_SHARED_MEMORY,
-            THREAD_PER_BLOCK=THREAD_PER_BLOCK,
-        )
-        return PixA_resamp
 
 
 class SpaceSFFT_NumpyFlow:
